requirements.txt
# =========================
# Web / API
# =========================
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0

# =========================
# LangChain / LangGraph Core
# =========================
langchain-core>=0.2.0
langchain-openai>=0.1.7
langgraph>=0.0.40

# =========================
# LLM Client (DeepSeek / OpenAI compatible)
# =========================
openai>=1.30.0

# =========================
# Utilities
# =========================
python-dotenv>=1.0.1

